+++ plugin/zxz-moe-bis/src/utils.ts
@@ -75,6 +75,31 @@ export function isBisInstalled(): Promise<void> {
     });
 }
 
+export function queryLocationOfBUILD(targetPath: string): Promise<string> {
+    return new Promise((resolve, reject) => {
+        vscode.commands
+            .executeCommand<string | undefined>("zxz-moe-bis.workspace", true)
+            .then((workspaceRoot) => {
+                promisify(executeBazelCommands)(
+                    ["query", targetPath, "--output=location"],
+                    workspaceRoot
+                )
+                    .then((stdout) => {
+                        const r = RegExp(/(.*\/BUILD(?:\.bazel)?\:\d+\:\d+)\:\s+.*rule.*/)
+                        if (r.test(stdout)) {
+                            resolve(stdout.match(r)![1]);
+                        } else {
+                            logger.error("Unexpect output: ", stdout);
+                            reject(new Error("no BUILD file found"));
+                        }
+                    })
+                    .catch((error) => {
+                        reject(error);
+                    });
+            });
+    });
+}
+
 export function touchBisBuild() {
     // Touch .bis/BUILD
     const task = new vscode.Task(